+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - new game data that is saved to an existing file does not overwrite other game data
 - make sure that new game url is unused
+- catch error when notification is invalid RDF
 
 ## [0.0.4] - 2018-11-10