+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+### [0.0.6] - 2018-11-12
+
+### Added
+
+- show modal when you do not have write permission for file
+
+### Fixed
+
+- generating new game url failed when file did not exist
+
 ### [0.0.5] - 2018-11-12
 
 ### Fixed
@@ -37,6 +47,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - use [semantic-chess](https://github.com/pheyvaer/semantic-chess-js), which replaces the local semantic chess code
 
+[0.0.6]: https://github.com/pheyvaer/solid-chess/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/pheyvaer/solid-chess/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/pheyvaer/solid-chess/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/pheyvaer/solid-chess/compare/v0.0.2...v0.0.3